Moved persistent history tracking and deduplication into separate files
